Normalize legacy finding details consistently
